Refactored auth modal state into react context.
