+++ components/Landing/Hero.tsx
@@ -1,6 +1,10 @@
 import DemoReceipt from './DemoReceipt';
+import { useAuthModalContext } from '@/context/authModalContext';
 
 export default function Hero() {
+  const [isOpen, setIsAuthModalOpen, view, setAuthModalView] =
+    useAuthModalContext();
+
   return (
     <section className="mx-auto max-w-2xl lg:max-w-6xl h-screen flex flex-col lg:flex-row items-center px-8 py-2 md:py-8">
       <div className="text-carbon-bronze basis-1/2 text-center lg:text-left">
@@ -16,6 +20,9 @@ export default function Hero() {
           <button
             type="button"
             className="bg-carbon-bronze text-carbon-gold rounded-2xl py-3 px-5 mt-6"
+            onClick={() => {
+              setIsAuthModalOpen(true), setAuthModalView('sign_up');
+            }}
           >
             Get started
           </button>
