+++ package.json
@@ -20,6 +20,7 @@
     "@supabase/auth-ui-shared": "^0.1.3",
     "@supabase/supabase-js": "^2.11.0",
     "classnames": "2.3.2",
+    "framer-motion": "^10.8.3",
     "next": "^13.2.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
