+++ pages/_document.tsx
@@ -4,7 +4,9 @@ class MyDocument extends Document {
   render() {
     return (
       <Html lang="en">
-        <Head />
+        <Head>
+          <link rel="stylesheet" href="https://use.typekit.net/zkr2bhm.css" />
+        </Head>
         <body className="loading">
           <Main />
           <NextScript />
